test(memlimit): add tests for new cgroup implementation
